+++ src/onegov.election_day/onegov/election_day/assets/js/d3.map.districts.js
@@ -59,13 +59,35 @@
 
 
         var districts;
-        var applyData = function(districts) {
+        var expats;
+        var applyData = function(data) {
             if (districts && scale) {
                 districts.attr('fill', function(d) {
-                    return d.value.counted ? scale(d.value.percentage) : 'url(#uncounted)';
+                    if (!isUndefined(d.value)) {
+                        if (d.value.counted) {
+                            return scale(d.value.percentage);
+                        }
+                        return 'url(#uncounted)';
+                    }
+                    return '#eee';
                 })
                 .attr('class', function(d) {
-                    return d.value.counted ? 'counted' : 'uncounted';
+                    if (!isUndefined(d.value)) {
+                        if (d.value.counted) {
+                            return 'counted';
+                        }
+                        return 'uncounted';
+                    }
+                    return 'extraneous';
+                });
+            }
+            if (expats && data && scale && ("" in data)) {
+                expats.attr('fill', function(d) {
+                    d.value = data[""];
+                    if (!isUndefined(d.value)) {
+                        return scale(d.value.percentage);
+                    }
+                    return '#eee';
                 });
             }
         };
@@ -135,7 +157,7 @@
                         var features = mapdata.objects.municipalities.geometries.filter(function(s) { return selected.has(s.id); });
                         return path(topojson.merge(mapdata, features));
                     });
-                applyData(districts);
+                applyData(data);
 
                 // Add the unused map data as a single block
                 extraneous = d3.set(extraneous);
@@ -186,16 +208,14 @@
                 // Add the expats
                 var bboxMap = svg[0][0].getBBox();
                 if ("" in data) {
-                    // the globe is 230px unscaled, we place it in the lower
-                    // left corner for now
-                    globe = svg.append('g')
+                    // the globe is 230px unscaled, we place it in the lower left corner for now
+                    var globe = svg.append('g')
                         .attr(
                             'transform',
                             'translate(0,' + Math.round(bboxMap.y + 3 / 4 * bboxMap.height) +  ')'
                         )
                         .property('__data__', {
                             'key': labelExpats,
-                            'value': data[""]
                           }
                         )
                         .append('g')
@@ -204,10 +224,9 @@
                             'scale(' + (bboxMap.width - bboxMap.x)/230/10 + ')'
                         );
 
-                    globe.append('g')
+                    expats = globe.append('g')
                         .append('path')
                         .attr('fill', 'white')
-                        .attr('fill', scale(data[""].percentage))
                         .attr('stroke', 'none')
                         .attr('d', "M 306.11308,163.17191 C 306.11308,224.93199 256.04569,274.99881 194.2941,274.99881 C 132.53683,274.99881 82.472286,224.93144 82.472286,163.17191 C 82.472286,101.41465 132.53683,51.352937 194.2941,51.352937 C 256.04569,51.352937 306.11308,101.41465 306.11308,163.17191 L 306.11308,163.17191 z ");
 
@@ -224,6 +243,9 @@
                             .on('click', tooltip.show);
                     }
                 }
+
+                // Apply data
+                applyData(data);
 
                 // Add the the legend (we need to up/downscale the elements)
                 var unitScale = d3.scale.linear()
@@ -326,7 +348,7 @@
         chart.update = function(url) {
             d3.json(url, function(data)  {
                 districts.data(d3.entries(data));
-                applyData(districts);
+                applyData(data);
             });
         };