+++ src/onegov.election_day/HISTORY.rst
@@ -8,6 +8,9 @@ Changelog
 - Updates the WabstiCExprt (proporz) import to the latest version.
   [msom]
 
+- Fixes expats in maps not getting updated on change.
+  [msom]
+
 3.8.2 (2019-01-21)
 ~~~~~~~~~~~~~~~~~~~